refactor(renderables): centralize shader logic into a new base class
Introduced `Object` abstract class in `renderables` to handle shader compilation logic, ensuring reusability.
Updated `Star` to inherit from `Object`.
Cleaned up `MyGLRenderer` by delegating shader responsibilities to `Object`.
